dlpearsonIII {smwrBase}	R Documentation

Log-Pearson Type III distribution

Description

Density, cumulative probability, quantiles, and random generation for the log-Pearson Type III distribution.

Usage

dlpearsonIII(x, meanlog = 0, sdlog = 1, skew = 0)

plpearsonIII(q, meanlog = 0, sdlog = 1, skew = 0)

qlpearsonIII(p, meanlog = 0, sdlog = 1, skew = 0)

rlpearsonIII(n, meanlog = 0, sdlog = 1, skew = 0)

Arguments

x, q
vector of quantiles.

meanlog
vector of means of the distribution of the log-transformed data.

sdlog
vector of standard deviation of the distribution of the log-transformed data.

skew
vector of skewness of the distribution of the log-transformed data.

p
vector of probabilities.

n
number of observations. If length(n) > 1, then the length is taken to be the number required.

Details

Elements of x, q, or p that are missing will result in missing values in the returned data.

Value

Either the density (dlpearsonIII), cumulative probability (plpearsonIII), quantile (qlpearsonIII), or random sample (rlpearsonIII) for the described distribution.

Note

The log-Pearson Type III distribution is used extensively in flood-frequency analysis in the United States.

See Also

dpearsonIII, dlnorm

Examples

## Simple examples
dlpearsonIII(c(.5, .75, .9), 1.5, .25, 0)
## compare to normal
qlnorm(c(.5, .75, .9), 1.5, .25)
## Make a skewed distribution
dlpearsonIII(c(.5, .75, .9), 1.5, .25, 0.25)

## Simple examples
plpearsonIII(c(.5, .75, .9), 1.5, .25, 0)
## compare to normal
qlnorm(c(.5, .75, .9), 1.5, .25)
## Make a skewed distribution
plpearsonIII(c(.5, .75, .9), 1.5, .25, 0.25)

## Simple examples
qlpearsonIII(c(.5, .75, .9), 1.5, .25, 0)
## compare to normal
qlnorm(c(.5, .75, .9), 1.5, .25)
## Make a skewed distribution
qlpearsonIII(c(.5, .75, .9), 1.5, .25, 0.25)

## Simple examples
rlpearsonIII(c(.5, .75, .9), 1.5, .25, 0)
## compare to normal
qlnorm(c(.5, .75, .9), 1.5, .25)
## Make a skewed distribution
rlpearsonIII(c(.5, .75, .9), 1.5, .25, 0.25)

[Package smwrBase version 1.1.5 Index]