Math-timeDay {smwrBase}	R Documentation

Mathematical Functions for timeDay objects

Description

No mathematical functions, such as log or exp are allowed on time-of-day data. An error results when trying to use any mathematical function on objects of class "timeDay."

Usage

## S4 method for signature 'timeDay'
Math(x)

Arguments

x
an object of class "timeDay."

[Package smwrBase version 1.1.5 Index]