dpearsonIII {smwrBase}	R Documentation

Pearson Type III distribution

Description

Density, cumulative probability, quantiles, and random generation for the Pearson Type III distribution.

Usage

dpearsonIII(x, mean = 0, sd = 1, skew = 0)

ppearsonIII(q, mean = 0, sd = 1, skew = 0)

qpearsonIII(p, mean = 0, sd = 1, skew = 0)

rpearsonIII(n, mean = 0, sd = 1, skew = 0)

Arguments

x, q
vector of quantiles. Missing values are permitted and result in corresponding missing values in the output.

mean
vector of means of the distribution of the data.

sd
vector of standard deviation of the distribution of the data.

skew
vector of skewness of the distribution of the data.

p
vector of probabilities.

n
number of observations. If length(n) > 1, then the length is taken to be the number required.

Details

Elements of x, q, or p that are missing will result in missing values in the returned data.

Value

Either the density (dpearsonIII), cumulative probability (ppearsonIII), quantile (qpearsonIII), or random sample (rpearsonIII) for the described distribution.

Note

The log-Pearson Type III distribution is used extensively in flood-frequency analysis in the United States. The Pearson Type III forms the basis for that distribution.

See Also

dlpearsonIII, dnorm

Examples

## Simple examples
dpearsonIII(c(.5, .75, .9), 1.5, .25, 0)
## compare to normal
qnorm(c(.5, .75, .9), 1.5, .25)
## Make a skewed distribution
dpearsonIII(c(.5, .75, .9), 1.5, .25, 0.25)
## Simple examples
ppearsonIII(c(.5, .75, .9), 1.5, .25, 0)
## compare to normal
qnorm(c(.5, .75, .9), 1.5, .25)
## Make a skewed distribution
ppearsonIII(c(.5, .75, .9), 1.5, .25, 0.25)
## Simple examples
qpearsonIII(c(.5, .75, .9), 1.5, .25, 0)
## compare to normal
qnorm(c(.5, .75, .9), 1.5, .25)
## Make a skewed distribution
qpearsonIII(c(.5, .75, .9), 1.5, .25, 0.25)
# Simple examples
rpearsonIII(c(.5, .75, .9), 1.5, .25, 0)

[Package smwrBase version 1.1.5 Index]