as.data.frame.timeDay {smwrBase}	R Documentation

Data Frames

Description

Convert an object to class "data.frame."

Usage

## S3 method for class 'timeDay'
as.data.frame(x, row.names = NULL, optional = FALSE, ...,
  nm = deparse(substitute(x)))

Arguments

x
the time-of-day object to be converted.

row.names
NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional
logical. If TRUE, setting row names and converting column names to syntactic names is optional.

...
not used, required for other methods.

nm
the column name to create for x.

Value

A data frame is created containing the data in x.

See Also

as.data.frame

[Package smwrBase version 1.1.5 Index]