as.timeDay {smwrBase}	R Documentation

Methods for Function as.timeDay

Description

Valid conversions for function as.timeDay.

Usage

as.timeDay(time, format)

## S4 method for signature 'timeDay,missing'
as.timeDay(time, format)

## S4 method for signature 'numeric,missing'
as.timeDay(time, format)

## S4 method for signature 'character,character'
as.timeDay(time, format)

## S4 method for signature 'character,missing'
as.timeDay(time, format)

Arguments

time
the time of day in character format.

format
The format for converting the time of day. See strptime for possible format specifiers.

Details

Inconsistent formats for time will result in an error. Missing values or empty strings in time will result in missing values in the output.

Examples

as.timeDay("10:00")
as.timeDay("3 PM", format="%I %p")

[Package smwrBase version 1.1.5 Index]