baseDay2decimal {smwrBase}	R Documentation

Base Day

Description

Computes the decimal time representation of the base day of the year.

Usage

baseDay2decimal(x)

Arguments

x
a vector of baseDay values, character, or factors of the form month abbreviation and day number, generally created from baseDay. Missing values are permitted and result in missing values in the output. Unmatched values also result in missing values in the output.

Value

A numeric value representing the base day.

See Also

baseDay

Examples

# The baseDay ordered by calendar year
bd.tmp <- baseDay(c("2000-02-29", "2000-03-01", "2001-03-01"), 
  numeric=FALSE)
baseDay2decimal(bd.tmp)
# ordered by water year, result should agree
bd.tmp <- baseDay(c("2000-02-29", "2000-03-01", "2001-03-01"), 
  numeric=FALSE, year="water")
baseDay2decimal(bd.tmp)

[Package smwrBase version 1.1.5 Index]