conc.meq {smwrBase}	R Documentation

Special Data

Description

Support function to supply data for functions within smwrBase.

Usage

conc.meq()

Value

A list containing necessary information for the function conc2meq.

Note

The user may choose to make local copies of the data with the same name (conc.meq) to be able to change or add to the list.

See Also

conc2meq

[Package smwrBase version 1.1.5 Index]