conc2meq {smwrBase}	R Documentation

Concentrations to Milliequivalents

Description

Convert concentrations in milligrams per liter (mg/L) to milli-equivalents per liter.

Usage

conc2meq(conc, constituent)

Arguments

conc
a numeric vector containing the concentration data in milligrams per liter.

constituent
the name of the constituent. Must be one of "aluminum," "ammonium," "bicarbonate," "bromide," "calcium," "carbonate," "chloride," "fluoride," "iron," "magnesium," "manganese," "nitrate as n," "nitrite as n," "phosphorus as p," "potassium," "sodium," "sulfate," or "sulfide." There must be enough characters in the name to uniquely identify the constituent. The case of the input name is ignored.

Value

Vector containing the milli-equivalent values. Missing values (NAs) are returned if the constituent name is invalid.

Note

The user must verify that the units of concentration are milligrams per liter. Only those constituents that are typically reported in milligrams per liter (rather than micrograms per liter) are provided in this function. Aluminum, iron, and manganese (and possibly sulfide) are sometimes reported in micrograms per liter. Such values should be divided by 1000.0 before using this function.

The conversion for iron assumes that the dissolved iron is iron II.

The conversion for phosphorus assumes that most of the phosphorus is divalent. The actual conversion for phosphorus is very dependent on pH.

The conversion factors are taken from table 9 (page 56) of Hem (1985). The available conversion factors are stored in the list created by conc.meq in smwrBase.

References

Hem, J.D., 1985, Study and interpretation of the chemical characteristics of natural water: U.S. Geological Survey Water-Supply Paper 2254, 263 p.

Examples

conc2meq(c(1,2,3), "Nitrate")
# should be: [1] 0.07139 0.14278 0.21417

[Package smwrBase version 1.1.5 Index]