dms2dd {smwrBase}	R Documentation

Decimal Degrees

Description

Convert data in degrees, minutes, and seconds format to decimal degrees.

Usage

dms2dd(x, minutes = NULL, seconds = 0, split = "")

Arguments

x
a character or numeric vector coded as degrees, minutes, and seconds or a numeric vector of degrees. The character string may contain a leading zero for values less than 100. Missing values are permitted and result in missing values in the output.

minutes
a vector of minutes. If supplied, then x is assumed to be a numeric vector of degrees. Missing values are permitted.

seconds
a vector of seconds. Assumed to be 0 if not supplied. Missing values are permitted.

split
the delimiter for x if x is character.

Value

A numeric vector of decimal degrees the same length as x. Missing values are returned wherever x, minutes, or seconds has a missing value.

Examples

dms2dd(983206) # using a numeric value
# should be [1] 98.535
dms2dd("0983206") # using a character value
# should be [1] 98.535
dms2dd(98, 32, 6) # using numeric values for degrees, minutes and seconds
# should be [1] 98.535
dms2dd("98:32", split=":") # Note seconds not included in text
# should be [1] 98.53333

[Package smwrBase version 1.1.5 Index]