exportRDB {smwrBase}	R Documentation

Export Data

Description

Exports a data frame to a text-based file.

Usage

exportRDB(x, file.name = "data.rdb", col.names = NULL, meta = FALSE,
  code.rule = 10)

Arguments

x
the data frame to be written.

file.name
a character string naming the file for output.

col.names
a vector of column names to use instead of the column names in x.

meta
a logical value indicating whether the header should include a metadata template for documentation or not.

code.rule
an integer value indicating how many unique numeric values should be included in the metadata template for cases where each distinct value has a descriptive meaning rather than a numeric meaning.

Details

Setting the meta argument to TRUE generates a header that contains a template that can be edited by the user to describe the contents of the file.

Value

The file name is returned. write.table, importRDB

[Package smwrBase version 1.1.5 Index]