makeMeta {smwrBase}	R Documentation

Metadata

Description

Create a template meta file for a CSV file. The meta file is used by importCSV to define column types.

Usage

makeMeta(file.name = "")

Arguments

file.name
the name of the CSV file, which may include the path.

Value

The name of the meta file that was created.

Note

The meta file that is created will only contain column types of character, numeric, integer, and logical. It may need to be edited by the user to redefine the column types actually needed for the data, for example columns of class "Date," "POSIXct," or "factor."

See Also

importCSV

[Package smwrBase version 1.1.5 Index]