more {smwrBase}	R Documentation

Display Data

Description

Display the contents of an object by pages.

Usage

more(x, n = 20L, ...)

Arguments

x
any valid object, generally a data frame, matrix, or table.

n
a positive integer indicating how many lines to print for a page.

...
additional arguments to be passed to methods for head or tail.

Value

The object x is returned invisibly. Sections of x of length n are displayed during the execution of the function.

Note

The function more is intended for interactive sessions. If used in a non-interactive session, it simply returns x invisibly.

Several keyboard commands can be used to view the contents of x. The function more will display n lines of x and wait for user input. Any of the following commands can be entered by the user; either upper- or lowercase letters are accepted.

q Quit

t Go to top of x

b Go to bottom of x

u Go up 1/2 page

p Go to previous page

d Go down 1/2 page

colName/pattern Search for pattern in the column named colName

h or ? Print help

any other letter Go down full page

Searching for a pattern in a column uses grep to search for the specified pattern in the character representation of the data in the column. This makes it possible to search columns that are not type character.

See Also

head, tail, grep

[Package smwrBase version 1.1.5 Index]