seqCollapse {smwrBase}	R Documentation

Collapse a Sequence

Description

Collapse a numeric sequence into a compact form that represents continuous ranges and discontinuous values.

Usage

seqCollapse(x, sequential = "-", skips = ", ")

Arguments

x
an integer vector, missing values and repeated values are permitted and removed before collapsing.

sequential
the separator for sequential values.

skips
the separator for gaps in the sequence.

Value

A character string that represents that data in x in a compact form. If x is empty, then "" is returned.

Note

This function is commonly used to express years in a compact form.

See Also

paste

Examples

# A single value
seqCollapse(1968)
# A single, continuous range of values
seqCollapse(1968:1992)
# A collection of continuous and individual values
seqCollapse(c(1968:1992, 1998, 2002, 2006:2012))

[Package smwrBase version 1.1.5 Index]