smwrBase-package {smwrBase}	R Documentation

Data import, export and manipulation functions

Description

This package has specialized functions for importing, managing, or manipulating hydrologic data.

Details

Package: 	 smwrBase

Type: 	 Package

Version: 	 1.1.2

Date: 	 2015-12-08

License: 	 File CC0

Depends: 	 methods,memoise,digest,lubridate

This package contains functions that import, manage, or manipulate hydrologic data and functions that apply specialized transformations used in hydrologic analyses amd modeling. A listing of the functions and their description is in the following table.

Function 	 Description

%cn% 	 Identify character strings that contain the specified pattern.

anomalies 	 Break down time-series data into long- and short-term deviations (anomalies)

and the high-frequency variation.

as.timeDay 	 Convert data to objects of class "timeDay."

baseDay 	 Computes the "base" day of the year, a reference value that can be used to

group days for the computation of summary statistics.

boxCox 	 Apply a Box-Cox power transformation.

coalesce 	 Merge a matrix or list of vectors selecting the first non-missing value.

conc.meq 	 A list containing necessary information for the function conc2meq.

conc2meq 	 Convert concentration in milligrams per liter to milli-equivalents per liter.

daysInMonth 	 The number of days in a month.

dectime 	 Convert dates and times to decimal time in years.

dectime2Date 	 Convert decimal time in years to an object of class "Date."

dlpearsonIII 	 The density of the log-Pearson Type III distribution.

dms2dd 	 Convert data in degrees, minutes, and seconds to decimal degrees.

dpearsonIII 	 The density of the Pearson Type III distribution.

eventLen 	 Compute the length or duration of an event.

eventNum 	 Compute the number of an event, identified by a TRUE value in a sequence.

eventSeq 	 Compute the sequence number within an event.

eventSeries 	 Create regular time-series data from recorded events.

exportCSV 	 Export a data frame to a comma-separated values file.

exportRDB 	 Export data to an ASCII relational-database file.

fillMissing 	 Interpolate missing values in a regular time-series of data.

fourier 	 Compute the Fourier series decomposition from date data.

group2row 	 Unstack data oriented in columns to rows of data.

hyperbolic 	 Apply a hyperbolic transformation.

hysteresis 	 Compute a basis for estimating hysteresis effect in some variable related to

the argument x.

IboxCox 	 Apply the inverse Box-Cox power transformation.

Ihyperbolic 	 Apply the inverse hyperbolic transformation.

importCSV 	 Import a data frame from a comma-separated values file.

importRDB 	 Import a data frame from an ASCII relational-database file.

index.coalesce 	 Return the index column number instead of the values for the first non-missing value.

isCharLike 	 Determine whether the data be treated like character data.

IsCurve 	 Apply the inverse s-curve transformation.

isDateLike 	 Deterimne whether the data can be treated like date data.

isGroupLike 	 Determine whether the data can be treated like grouping data.

isNumberLike 	 Determine whether the data can be treated as numeric data.

makeMeta 	 Create a template meta file for a comma-separated values file.

mergeNearest 	 Merge two datasets by the nearest date and time.

mergeQ 	 Merge flow data with water-quality data.

miss2na 	 Convert a coded missing value to NA.

more 	 Display the contents of an object by pages.

movingAve 	 Compute the moving average in regular time-series data.

movingDiff 	 Compute the moving difference in regular time-series data.

na2miss 	 Convert NA to a coded missing value.

peaks 	 Compute the indices of peaks in time-series data.

pick 	 Select a value based on the value of a logical, integer, or character reference value.

plpearsonIII 	 Compute the cumulative probability of the log-Pearson Type III distribution.

ppearsonIII 	 Compute the cumulative probability of the Pearson Type III distribution.

qlpearsonIII 	 Compute the quantile of the log-Pearson Type III distribution.

qpearsonIII 	 Compute the quantile of the Pearson Type III distribution.

quadratic 	 Compute a basis for an orthogonal second-order polynomial.

readList 	 Import data arranged on lines into a list.

recode 	 Recode distinct values.

regularSeries 	 Put data collected at arbitrary times into a regular time series.

rlpearsonIII 	 Compute the random variates of the log-Pearson Type III distribution.

rpearsonIII 	 Compute the random variates of the Pearson Type III distribution.

sCurve 	 Apply the s-curve transformation.

scaleRng 	 Scale data to a specified range.

screenData 	 Screen data for missing values or gaps.

seasons 	 Create seasonal categories from dates.

seqCollapse 	 Collapse a sequence of integers to a compact character string.

setFileType 	 Support function to manage file suffixes.

setTZ 	 Set the time zone information for dates and times.

shiftData 	 Shift time-series data forward or backward.

sumComposition 	 Compute the percentages of data within a matrix.

timeDay 	 Various methods for manipulating time-of-day data, including conversion to and

from character, addition, and others.

untable 	 Expand a 2-dimensional table into the raw values.

waterYear 	 Compute the water year of date data. The water year ends on September 30 of the year.

whichRowCol 	 Identify the row and column indexes for TRUE values in a logical matrix.

Author(s)

Dave Lorenz

References

Lorenz, D.L., 2015, smwrBase—an R package for managing hydrologic data, version 1.1.1: U.S. Geological Survey Open-File Report 2015–1202, 7 p.

See Also

smwrData

[Package smwrBase version 1.1.5 Index]