timeDay-class {smwrBase}	R Documentation

Time of Day

Description

Class "timeDay" describes the time of day without any reference to the date. The data are stored as seconds since midnight.

Slots

time

the time of day in seconds since midnight.

format

a character string indicating the format to display the time of day.

Objects from the Class

Objects can be created by calls of the form as.timeDay(time, format).

Examples

showClass("timeDay")

[Package smwrBase version 1.1.5 Index]